ai_agent/requirements.txt
fastapi
uvicorn
openai
supabase
python-dotenv
pydantic
requests
